# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Penta Chatbot API",
        "version": "1.0"
    },
    "paths": {
        "/api/chat": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Handle AI chat",
                "description": "Handle incoming chat messages and get responses from the AI.",
                "operationId": "40b2b26fa6f1d97fded6b2c2b04445fa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "messages": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chat response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chat-logs": {
            "get": {
                "tags": [
                    "Chat Logs"
                ],
                "summary": "Get chat logs",
                "description": "Get the chat logs (Admin only).",
                "operationId": "c510328ade03e3a9da7274ad246c9215",
                "responses": {
                    "200": {
                        "description": "List of chat logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChatLog"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chat-logs/{id}": {
            "delete": {
                "tags": [
                    "Chat Logs"
                ],
                "summary": "Delete chat log",
                "description": "Delete a chat log (Admin only).",
                "operationId": "5c0c69c541af2ea80022e1de25291c82",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the chat log to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat log deleted"
                    },
                    "404": {
                        "description": "Chat log not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Get all users.",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a user",
                "description": "Create a new user.",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "permission_group"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User email address",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "description": "User phone number",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password for the user",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "permission_group": {
                                        "description": "Permission group of the user",
                                        "type": "string",
                                        "enum": [
                                            "normal",
                                            "moderator",
                                            "admin",
                                            "superadmin",
                                            "developer"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a user by ID",
                "description": "Get a single user.",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user",
                "description": "Update an existing user.",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User email address",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "description": "User phone number",
                                        "type": "string"
                                    },
                                    "permission_group": {
                                        "description": "Permission group of the user",
                                        "type": "string",
                                        "enum": [
                                            "normal",
                                            "moderator",
                                            "admin",
                                            "superadmin",
                                            "developer"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "description": "Delete a user.",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ChatLog": {
                "title": "Chat Log",
                "description": "Chat log resource",
                "properties": {
                    "id": {
                        "description": "Chat log ID",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID of the user who sent the chat",
                        "type": "integer"
                    },
                    "message": {
                        "description": "Chat message",
                        "type": "string"
                    },
                    "response": {
                        "description": "AI response",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Timestamp of when the chat was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp of the last update",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User resource",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "User email address",
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "description": "User phone number",
                        "type": "string"
                    },
                    "permission_group": {
                        "description": "Permission group of the user",
                        "type": "string",
                        "enum": [
                            "normal",
                            "moderator",
                            "admin",
                            "superadmin",
                            "developer"
                        ]
                    },
                    "created_at": {
                        "description": "Timestamp when the user was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the user was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Chat",
            "description": "Chat"
        },
        {
            "name": "Chat Logs",
            "description": "Chat Logs"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}